+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/internal/storage/db/SQLVariants.java
@@ -2,4 +2,6 @@ package com.nitorcreations.nflow.engine.internal.storage.db;
 
 public interface SQLVariants {
   String currentTimePlusSeconds(int seconds);
+
+  boolean hasUpdateReturning();
 }